switch over to toy models
